Refactor input-and-results code to not be scattered accross files. Fix onDecode with string
