+++ src/lib/components/schedule/schedule-input-payload.svelte
@@ -20,6 +20,7 @@
   export let payloads: Payloads;
   export let showEditActions: boolean = false;
 
+  console.log('Payloads: ', payloads);
   let initialInput = '';
   let initialEncoding: PayloadInputEncoding = 'json/plain';
   let initialMessageType = '';
