+++ src/lib/components/schedule/schedule-input.svelte
@@ -1,17 +1,17 @@
 <script lang="ts">
   import Panel from '$lib/components/panel.svelte';
+  import InputAndResultsPayload from '$lib/components/workflow/input-and-results-payload.svelte';
   import { translate } from '$lib/i18n/translate';
   import type { Payloads } from '$lib/types';
   import { stringifyWithBigInt } from '$lib/utilities/parse-with-big-int';
-
-  import InputAndResults from '../workflow/input-and-results.svelte';
 
   export let input: Payloads;
 </script>
 
 <Panel>
-  <InputAndResults
+  <InputAndResultsPayload
     title={translate('schedules.schedule-input')}
     content={stringifyWithBigInt(input)}
+    isRunning={false}
   />
 </Panel>
