+++ src/lib/components/workflow/input-and-results-payload.svelte
@@ -2,31 +2,58 @@
   import CodeBlock from '$lib/holocene/code-block.svelte';
   import { translate } from '$lib/i18n/translate';
   import { minimizeEventView } from '$lib/stores/event-view';
+  import type { Payload } from '$lib/types';
+  import type { PotentiallyDecodable } from '$lib/utilities/decode-payload';
   import {
     parseWithBigInt,
     stringifyWithBigInt,
   } from '$lib/utilities/parse-with-big-int';
 
-  import {
-    getPayloads,
-    parseContent,
-    parsePayloads,
-  } from './input-and-result-payloads';
   import PayloadDecoder from '../event/payload-decoder.svelte';
 
-  export let title: string;
-  export let content: string;
-  export let isRunning: boolean;
+  type Props = {
+    title: string;
+    content: string;
+    isRunning: boolean;
+  };
 
-  $: parsedContent = parseContent(content);
-  $: payloads = getPayloads(parsedContent);
-  $: showParsedContent = payloads.length > 0;
+  let { title, content = '', isRunning }: Props = $props();
+
+  const parseContent = (c: string): PotentiallyDecodable | undefined => {
+    try {
+      return parseWithBigInt(c);
+    } catch {
+      return undefined;
+    }
+  };
+
+  const parsePayloads = (c: string): unknown[] => {
+    try {
+      const data = parseWithBigInt(c);
+      return Array.isArray(data) ? data : [data];
+    } catch {
+      return [];
+    }
+  };
+
+  const getPayloads = (value: PotentiallyDecodable | undefined): Payload[] => {
+    try {
+      const payloads = value?.payloads;
+      return Array.isArray(payloads) ? payloads : [];
+    } catch {
+      return [];
+    }
+  };
+
+  const parsedContent = $derived(parseContent(content));
+  const payloads = $derived(getPayloads(parsedContent));
+  const showParsedContent = $derived(payloads.length > 0);
 </script>
 
-<div class="flex w-full grow flex-col gap-2 {$$restProps.class}">
-  <h3 class="flex items-center gap-2">
+<div class="flex w-full grow flex-col gap-2">
+  <h2 class="flex items-center gap-2">
     {title}
-  </h3>
+  </h2>
   {#if content}
     {#key $minimizeEventView}
       {#if showParsedContent}
